% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cellPCT}
\alias{cellPCT}
\title{cellPCT: Multi-visualization for single-cell RNA-seq cell type ratio data.}
\description{
This package provide multiple visualizations for scRNA-seq cell type ratio.
Including sankey plot, stackbar plot, stackline plot, multidonut plot, singledonut plot,
circle plot, rose plot, treemap plot, radar plot and milticircle plot.
}
\details{
Each plot type can choose multi parameters to fit users' demands. Like label showing, coord flip,
hide cell type which ratio less assigned value, order the group and cell type for plotting, et al.
More details please see the description of 'plot_xx' functions.
}
