% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle.R
\name{circle_plot}
\alias{circle_plot}
\title{circle_plot function}
\usage{
circle_plot(data, lab)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
circle_p A ggplot2 object containing the plot.
}
\description{
circle_plot function
}
