% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example_data for cellPCT multiplots}
\format{
A data.frame with 4 columns and 10 cell types:
\describe{
  \item{Col1}{sample1}
  \item{Col2}{sample2}
  \item{Col3}{sample3}
  \item{Col4}{sample4}
}
}
\usage{
data(example_data)
}
\description{
Collection 4 samples and 10 cell types
This data set is the example for user's data preparetion
}
\keyword{datasets}
