% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multicircle.R
\name{multicircle_plot}
\alias{multicircle_plot}
\title{multicircle_plot function}
\usage{
multicircle_plot(plot_data, outdir)
}
\arguments{
\item{outdir}{output dir for multicircle_plot Eg. '/path/to/save/plot', default is './'}

\item{data}{input data, Seurat obj or data.frame can be use.}
}
\description{
multicircle_plot function
}
