% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidonut.R
\name{mutidonut_plot}
\alias{mutidonut_plot}
\title{mutidonut_plot function.}
\usage{
mutidonut_plot(data, order_sample, lab)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{order_sample}{reorder the samples.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
mutidonut_p A ggplot2 object containing the plot.
}
\description{
mutidonut_plot function.
}
