% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Cell type ration visualization in all plot with default paraments and save as png and pdf.}
\usage{
plot_all(input, outdir, group_by = c("sample"), cell_type = c("cluster.type"))
}
\arguments{
\item{input}{the input data rds,txt,csv can be use}

\item{outdir}{the output files' path}

\item{cell_type}{the colname of cell in input data}

\item{sample}{the colname of sample in input data}
}
\description{
Cell type ration visualization in all plot with default paraments and save as png and pdf.
}
\examples{
plot_all(input=input, outdir=outdir,group_by='sample',cell_type='cluster.type')
}
