% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multicircle.R
\name{plot_multicircle}
\alias{plot_multicircle}
\title{Cell type ration visualization in multicircle plot}
\usage{
plot_multicircle(
  input,
  group_by = NULL,
  cell_type = NULL,
  sub_sample = "all",
  sub_type = "all",
  order_type = "default",
  outdir = "."
)
}
\arguments{
\item{input}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_type}{reorder the cell type.}

\item{outdir}{output dir for multicircle_plot Eg. '/path/to/save/plot', default is './'}
}
\description{
Cell type ration visualization in multicircle plot
}
\examples{
plot_multicircle(input, group_by = 'sample',cell_type = 'cluster.type')
}
