% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multidonut.R
\name{plot_mutidonut}
\alias{plot_mutidonut}
\title{Cell type ration visualization in multidonut plot}
\usage{
plot_mutidonut(
  input,
  group_by = NULL,
  cell_type = NULL,
  sub_sample = "all",
  sub_type = "all",
  order_sample = "default",
  order_type = "default",
  min_percent = 0,
  lab = T
)
}
\arguments{
\item{input}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_sample}{reorder the samples.}

\item{order_type}{reorder the cell type.}

\item{min_percent}{the minimum percent. use to combine some cell type with low ratio, default is 0.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
mutidonut A ggplot2 object containing the plot.
}
\description{
Cell type ration visualization in multidonut plot
}
\examples{
plot_mutidonut(input, group_by = 'sample',cell_type = 'cluster.type')
}
