% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rose.R
\name{plot_rose}
\alias{plot_rose}
\title{Cell type ration visualization in rose plot}
\usage{
plot_rose(
  input,
  group_by = NULL,
  cell_type = NULL,
  sub_sample = "all",
  sub_type = "all",
  order_type = "default",
  lab = T
)
}
\arguments{
\item{input}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_type}{reorder the cell type.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
rose A list contain ggplot2 objects
}
\description{
Cell type ration visualization in rose plot
}
\examples{
plot_rose(input, group_by = 'sample',cell_type = 'cluster.type')
}
