% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Cell type ration visualization in sankey plot}
\usage{
plot_sankey(
  input,
  group_by = NULL,
  cell_type = NULL,
  sub_sample = "all",
  sub_type = "all",
  order_sample = "default",
  order_type = "default",
  coord_flip = F
)
}
\arguments{
\item{input}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_sample}{reorder the samples.}

\item{order_type}{reorder the cell type.}

\item{coord_flip}{whether to flip the coord, default is FALSE.}
}
\value{
sankey A ggplot2 object containing the plot.
}
\description{
Cell type ration visualization in sankey plot
}
\examples{
plot_sankey(input, group_by = 'sample',cell_type = 'cluster.type')
}
