% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackbar.R
\name{plot_stackbar}
\alias{plot_stackbar}
\title{Cell type ration visualization in bar plot}
\usage{
plot_stackbar(
  input,
  group_by = NULL,
  cell_type = NULL,
  sub_sample = "all",
  sub_type = "all",
  order_sample = "default",
  order_type = "default",
  group = FALSE
)
}
\arguments{
\item{input}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_sample}{reorder the samples.}

\item{order_type}{reorder the cell type.}

\item{group}{whether use group to use as legend, default is FALSE.}
}
\value{
stackbar A ggplot2 object containing the plot.
}
\description{
Cell type ration visualization in bar plot
}
\examples{
plot_stackbar(input, group_by = 'sample',cell_type = 'cluster.type')
}
