% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare plot data for multi-visualization functions.}
\usage{
prepare_data(data, group_by, cell_type, sub_sample, sub_type, order_type)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{group_by}{colname of Seurat metadata, which represent the groups or samples.}

\item{cell_type}{colname of Seurat metadata, which represent the cell type.}

\item{sub_sample}{select part of the groups for plot, default is 'all'.}

\item{sub_type}{select part of the cell types for plot, default is 'all'.}

\item{order_type}{reorder the cell type.}
}
\value{
sub_prepared_cell
}
\description{
Prepare plot data for multi-visualization functions.
}
