% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_multicircle}
\alias{prepare_multicircle}
\title{Prepare data for multicircle plot}
\usage{
prepare_multicircle(prepared)
}
\arguments{
\item{prepared}{the return value of prepare_data.}
}
\value{
multicircle_data
}
\description{
Prepare data for multicircle plot
}
