% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_radar}
\alias{prepare_radar}
\title{Prepare data for radar plot}
\usage{
prepare_radar(prepared)
}
\arguments{
\item{prepared}{the return value of prepare_data.}
}
\value{
radar_data
}
\description{
Prepare data for radar plot
}
