% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{radar_plot}
\alias{radar_plot}
\title{radar_plot function}
\usage{
radar_plot(data)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}
}
\value{
radar_p A ggplot2 object containing the plot.
}
\description{
radar_plot function
}
