% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rose.R
\name{rose_plot}
\alias{rose_plot}
\title{rose_plot function.}
\usage{
rose_plot(data, lab)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
rose_p A list contain ggplot2 objects
}
\description{
rose_plot function.
}
