% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{sankey_plot}
\alias{sankey_plot}
\title{sankey_plot function.}
\usage{
sankey_plot(data, order_sample, coord_flip)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{order_sample}{reorder the samples.}

\item{coord_flip}{whether to flip the coord, default is FALSE.}
}
\value{
sankey_p A ggplot2 object containing the plot.
}
\description{
sankey_plot function.
}
