% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{show_others}
\alias{show_others}
\title{Select part of the Prepare plot data for multi-visualization functions.}
\usage{
show_others(sample_ratio, min_percent)
}
\arguments{
\item{sample_ratio}{the return value of prepare_data.}

\item{min_percent}{the minimum percent. use to combine some cell type with low ratio,default is 0.}
}
\value{
sample_new
}
\description{
Select part of the Prepare plot data for multi-visualization functions.
}
