% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singledonut.R
\name{singledonut_plot}
\alias{singledonut_plot}
\title{singledonut_plot function}
\usage{
singledonut_plot(data, order_sample, lab)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{order_sample}{reorder the samples.}

\item{lab}{whether to show the label on the plot, default is TRUE.}
}
\value{
singledonut_p A ggplot2 object containing the plot.
}
\description{
singledonut_plot function
}
