% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackbar.R
\name{stackbar_plot}
\alias{stackbar_plot}
\title{stackbar_plot function.}
\usage{
stackbar_plot(data, order_sample, group)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{order_sample}{reorder the samples.}

\item{group}{whether use group to use as legend, default is FALSE.}
}
\value{
stackbar_p A ggplot2 object containing the plot.
}
\description{
stackbar_plot function.
}
