% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackline.R
\name{stackline_plot}
\alias{stackline_plot}
\title{stackline_plot function}
\usage{
stackline_plot(data, order_sample)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}

\item{order_sample}{reorder the samples.}
}
\value{
stackline_p A ggplot2 object containing the plot.
}
\description{
stackline_plot function
}
