% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemap.R
\name{treemap_plot}
\alias{treemap_plot}
\title{treemap_plot function}
\usage{
treemap_plot(data)
}
\arguments{
\item{data}{input data, Seurat obj or data.frame can be use.}
}
\value{
treemap_p A list contain ggplot2 objects
}
\description{
treemap_plot function
}
